Move favicons to it’s own component
