+++ src/components/Layout/favicon.js
@@ -1,0 +1,21 @@
+import React from 'react'
+import Helmet from 'react-helmet'
+
+const Favicon = () => {
+  return(
+    <Helmet>
+      <link rel="shortcut icon" type="image/vnd.microsoft.icon" href="https://magento.com/favicon.ico" />
+      <link rel="apple-touch-icon" type="image/png" href="https://magento.com/apple-touch-icon.png" />
+      <link rel="apple-touch-icon" type="image/png" sizes="57x57" href="https://magento.com/apple-touch-icon-57x57.png" />
+      <link rel="apple-touch-icon" type="image/png" sizes="72x72" href="https://magento.com/apple-touch-icon-72x72.png" />
+      <link rel="apple-touch-icon" type="image/png" sizes="76x76" href="https://magento.com/apple-touch-icon-76x76.png" />
+      <link rel="apple-touch-icon" type="image/png" sizes="114x114" href="https://magento.com/apple-touch-icon-114x114.png" />
+      <link rel="apple-touch-icon" type="image/png" sizes="120x120" href="https://magento.com/apple-touch-icon-120x120.png" />
+      <link rel="apple-touch-icon" type="image/png" sizes="144x144" href="https://magento.com/apple-touch-icon-144x144.png" />
+      <link rel="apple-touch-icon" type="image/png" sizes="152x152" href="https://magento.com/apple-touch-icon-152x152.png" />
+      <link rel="apple-touch-icon" type="image/png" sizes="180x180" href="https://magento.com/apple-touch-icon-180x180.png" />
+    </Helmet>
+  )
+}
+
+export default Favicon
